+++ install-stubs/Orchid/Composers/SystemMenuComposer.php
@@ -45,7 +45,7 @@ class SystemMenuComposer
                     ->route('platform.systems.users')
                     ->permission('platform.systems.users')
                     ->sort(1000)
-                    ->groupName(__('All registered users'))
+                    ->title(__('All registered users'))
             )
             ->add('Auth',
                 ItemMenu::label(__('Roles'))
@@ -53,7 +53,7 @@ class SystemMenuComposer
                     ->route('platform.systems.roles')
                     ->permission('platform.systems.roles')
                     ->sort(1000)
-                    ->groupName(__('A Role defines a set of tasks a user assigned the role is allowed to perform. '))
+                    ->title(__('A Role defines a set of tasks a user assigned the role is allowed to perform. '))
             )
             ->add('CMS',
                 ItemMenu::label(__('Categories'))
@@ -61,7 +61,7 @@ class SystemMenuComposer
                     ->route('platform.systems.category')
                     ->permission('platform.systems.category')
                     ->sort(1000)
-                    ->groupName(__('Sort entries into groups of posts on a given topic. This helps the user to find the necessary information on the site.'))
+                    ->title(__('Sort entries into groups of posts on a given topic. This helps the user to find the necessary information on the site.'))
             )
             ->add('CMS',
                 ItemMenu::label(__('Comments'))
@@ -69,7 +69,7 @@ class SystemMenuComposer
                     ->route('platform.systems.comments')
                     ->permission('platform.systems.comments')
                     ->sort(1000)
-                    ->groupName(__("Comments allow your website's visitors to have a discussion with you and each other."))
+                    ->title(__("Comments allow your website's visitors to have a discussion with you and each other."))
                     ->badge(function () {
                         return \Orchid\Press\Models\Comment::where('approved', 0)->count() ?: null;
                     })